fix: remove unnecessary cast for Linux clippy
Fixes clippy unnecessary_cast on statvfs free space calculation (Linux/WSL).

+++ src-tauri/src/updater/helper.rs
@@ -2084,7 +2084,7 @@ fn available_disk_space(path: &Path) -> Option<u64> {
         return None;
     }
     let stat = unsafe { stat.assume_init() };
-    Some((stat.f_bavail as u64).saturating_mul(stat.f_frsize))
+    Some(stat.f_bavail.saturating_mul(stat.f_frsize))
 }
 
 #[cfg(not(any(unix, target_os = "windows")))]
